+++ example/lib/src/another_app/home/gridview/images/controller/cat_controller.dart
@@ -8,6 +8,6 @@ import 'package:example/src/controller.dart';
 class CatController extends InheritController {
   ///
   factory CatController() => _this ??= CatController._();
-  CatController._() : super();
+  CatController._();
   static CatController? _this;
 }