+++ example/lib/src/another_app/home/gridview/images/controller/dog_controller.dart
@@ -8,6 +8,6 @@ import 'package:example/src/controller.dart';
 class DogController extends InheritController {
   ///
   factory DogController() => _this ??= DogController._();
-  DogController._() : super();
+  DogController._();
   static DogController? _this;
 }